requirements.txt
# Core NLP and ML libraries
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
nltk>=3.8.0
spacy>=3.6.0
transformers>=4.30.0
torch>=2.0.0

# Conversational AI frameworks
rasa>=3.6.0
rasa-sdk>=3.6.0

# LLM APIs
openai>=1.0.0
anthropic>=0.3.0
google-generativeai>=0.3.0

# Web frameworks
flask>=2.3.0
fastapi>=0.100.0
uvicorn>=0.23.0
websockets>=11.0

# Database
sqlalchemy>=2.0.0
pymongo>=4.4.0
redis>=4.6.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
locust>=2.15.0

# Analytics and monitoring
prometheus-client>=0.17.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
pyyaml>=6.0
